tg-tag: when dropping @{0} avoid deleting the ref
Normally when dropping @{0} on a non-symbolic ref, the ref itself
will be updated to point to the new @{0} log entry.

However, if @{0} was the last log entry that will delete the ref
itself even though the command was "drop" and not "delete".

Change this so that using "drop" does not delete the ref, but
just the log entry.

The command to delete the ref entirely is "delete", not "drop" and
having a "delete" occur as a side-effect of a "drop" can be a bit
unexpected.

Nevertheless, even with these changes, if @{0} is dropped and it's
not the only entry in the reflog for a non-symbolic ref, but all
the other entries have invalid hash values (perhaps --reference was
used in repository creation and those objects have since been garbage
collected), the ref will end up being removed anyway.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>