tg-tag.sh: preserve ref when dropping @{0} reflog entry
Enhance the code that attempts to preserve the ref itself when
dropping the @{0} reflog entry of a non-symbolic ref that is the
last entry in the reflog by first performing a "--stale-fix" on
that reflog.

Previously it was possible the ref itself would still be affected
if the reflog contained more than one entry but all the other
entries pointed to invalid hash values.

Avoid this possibility by first running "--stale-fix" on the
reflog before allowing "--updateref" to be used.

There's still a miniscule possibility that a simultaneous garbage
collection could render the other entries invalid after the
"--stale-fix" validates them but before the "--updateref" operation
occurs thereby resulting in the ref ultimately being removed anyway.

In the general case, simultaneous garbage collection is not safe
while other operations are being performed on the repository and
this is just another example of one of the race conditions that
can occur.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>