tg-tag: avoid --stale-fix if possible when --drop'ing
The `tg tag --drop` command makes every attempt to preserve the
existing ref while affecting only the reflog itself.

However, if the @{0} entry is dropped, the ref will be updated
provided it's not a symbolic ref AND there's something valid to
update it to (i.e. @{0} wasn't the last entry in the reflog).

When using the Git alternates mechanism (via --shared and/or
--reference options for example), reflog entries can easily become
invalid AKA "stale".

In an effort to avoid losing a ref when doing `tg tag --drop ...@{0}`
care is taken to make sure the new ref value is actually valid and
not a "stale" entry.

This was being accomplished via "--stale-fix".  However, the
"--stale-fix" operation can be rather slow and expensive even when
operating on just a single ref name.

Compensate for this by short circuiting some common cases:

  1) There are less than two entries in the reflog

     In this case the ref itself will not be updated so there's
     no need to do a "--stale-fix" for any reason.

  2) The @{1} entry in the log is actually valid

     This can be checked much more quickly than running the
     "--stale-fix" command.  If the @{1} entry is valid there's
     no need to do any additional checking.

This will handle the vast majority of the cases and always covers
the case where there are no stale entries present in the reflog.

It's also generally much faster as a result.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>