tg-tag: avoid unreliable --stale-fix entirely
Besides the fact that --stale-fix can be very slow, it turns out
that it also does not always find and fix all the stale entries!

Instead just expand on the previous fast check by turning that into
a pseudo stale fix that removes a stale ...@{1} if it's found to
be broken and ...@{0} of a non-symbolic ref is being dropped.  Then
repeating that in a loop until a non-stale ...@{1} is found or all
the stale entries have been dropped.

This will still be as fast as the previous quick check as the loop
stops as soon as the first valid ...@{1} is found.  It will still
be faster than using --stale-fix (if that actually always worked)
except perhaps in the case of a reflog that consists of an outlandishly
large number of stale entries.  Normal Git operations do not create
stale reflog entries, the checks are a "just in case" fallback that
are expected to be used only very rarely.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>