tg-tag.sh: properly --clear reflogs
When clearing a reflog to a single entry, the "from" value
must be reset to all "0"s to avoid showing a straggler when
using `git log --walk-reflogs`.

Switch from sed to awk thereby allowing this to be done easily
without needing to run anymore external commands.

Also take this opportunity to add a missing `2>/dev/null`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>