tg-tag.sh: carefully resolve full ref names
Although this command is available:

    git rev-parse --revs-only --symbolic-full-name <refname>

It has some issues:

 1) It does not support a --no-deref option and always follows symrefs
 2) It fails to output anything when the name is ambiguous

The problem with (1) is that the tg tag operations need to operate
on the original ref, not anything it may point to if it's a symref.

The problem with (2) is that something like "origin" which should be
resolved according to the rules to "refs/remotes/origin/HEAD" (if that
exists) fails because it mistakenly thinks that there's an ambiguity
between "refs/remotes/origin" and "refs/remotes/origin/HEAD" even
though `git rev-parse --verify` will resolve it correctly.

Instead implement our own v_resolve_full_name function that uses a
combination of `git check-ref-format` and `git rev-parse --verify`
along with the list of interpretations to be tried from the list
in `git help revisions` to come up with the correct answer.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>