tg-tag: detach HEAD before deleting its symref
Just like the other operations that do the same make sure that when
`tg tag --delete <ref>` removes the ref that a symbolic HEAD points
to that HEAD first gets detached to that revision.

The alternatives are to a) just delete the target ref or b) deny
the operation.  Doing (a) leaves the working tree in a very ugly
state as it looks like it's about to commit a new revision on an
unborn branch.  On the other hand (b) can be quite frustrating if
the user wants to remove that ref and yet the "helpful" software
just refuses to do it (such as `git branch --force --delete master`
when HEAD is a symref to refs/heads/master).

Detaching HEAD first prevents any loss and allows the operation to
succeed without unnecessary frustration while preventing the old
behavior that left HEAD pointing to an unborn branch in that case.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>