tg.sh: accept --paginate as --pager alias
Git incredibly accepts `--no-pager` to disable any pager, `-p` to
force a pager, but not `--pager` to force a pager!

Instead it wants `--paginate` if one has pedantic tendencies.

But, notwithstanding, it will NOT accept `--no-paginate` as an alias
for for `--no-pager`!

Grrrrr.  Accomodate this bizarre state of affairs by also accepting
`--paginate` as an alias for `--pager` to avoid frustration.

However, since Git does NOT accept `--no-paginate` do not capitulate
to that.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>