tg-push: check for out-of-dateness before pushing
Do not allow out-of-date TopGit branches to be pushed unless the
`--allow-outdated` option has been given.

Additionally, support a `topgit.pushRemote` setting to better support
a 3-way Git workflow where changes are fetched from one upstream
remote (`topgit.remote`), worked on locally and then pushed to a
different remote (`topgit.pushRemote`).

Without `--allow-outdated` branches to be pushed will also be checked
for out-of-dateness against the upstream remote (`topgit.remote`)
if it's set (or has been given via a global `-r <remote>` option)
in addition to the local refs, but will NOT be checked against the
remote being pushed to (unless it happens to be the same remote).

And while in there making changes, add a few minor tidy ups and
option validation checks as well.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>