tg-info.sh: omit ignored dependencies
Annihilated branches do not logically have dependencies.

However, if they've been created by an older version of TopGit or
created (incorrectly) manually, a non-empty .topdeps file could be
present in the branch even though it's been annihilated.

Do not show dependencies in these cases.  They will be totally
ignored while recursing for updates or other activities and could
be misleading.  Since we already know the type of branch we are
showing it's easy to skip the display in this case.

In addition, skip the display for new branches being created as
well unless the -i or -w option has been used because they are
currently in the same state.  We also show dependencies for
annihilated/bare branches (if there are any) when -i or -w has been
used because that's what -i and -w will do for other tg commands.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>