tg.sh: require Git 1.9.0 or later
When TopGit 0.19.7 added code to keep the .topfiles out of the bases
and guarantee proper .topdeps merging from remote branches, it
started using the "exclude" pathspec magic.

Unfortunately this pathspec magic (see `git help glossary` the
pathspec entry), was not introduced until Git version 1.9.0.
(It's also known as "negative pathspec syntax".)

Update the minimum required Git version from 1.8.5 to 1.9.0 to
reflect this (there's no simple alternative to this functionality).

Git 1.8.5 was released on 2013-11-27 and Git 1.9.0 on 2014-02-14.

The effect of this change is to require a Git that is almost 3
months more recent than previously required, but still over 4 years
older than this commit itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>