tg.sh: omit --create-reflog option prior to Git version 2.10.0
When creating a non-read-only wayback directory, the --create-reflog
option is normally passed for all "wayback" ref changes in order to
force a ref log to be created for them in the wayback directory.

This is mostly fluff as there will only be the single entry.  It's
mainly intended to facilitate tracking changes that are made by the
user in that directory later on.

Git will create a ref log automatically for branches unless ref
logs have been explicitly disabled.  Adding the option provides
additional logs that may or may not actually be useful to the user.

Unfortunately the --create-reflog option only works with the
`git update-ref` command when running Git version 2.10.0 or later.

Therefore simply omit the option when using an earlier version of
Git.  Everything still works except that non-branches will not
automatically get an initial ref log in a non-read-only wayback
directory.  TopGit itself will force a ref log to be added for
certain opreations if ref logs have been enabled.  As a result
there should be almost no detectable differences to the end user.

Non-wayback operations are unaffected by this change as well
as read-only wayback operations.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>