topgit_deps.awk: clarify and refine exclbr/inclbr operation
Refine the operation of exclbr and inclbr to operate on the
"edges" of the dependency graph.

While inclbr still includes edges (leaving empty implicitly
includes all edges) it does so if *either* end appears in
inclbr.

While exclbr still excludes edges it does so if *either*
end appears in exclbr and trumps inclbr in the process.

With this change to operate on edges instead of the somewhat
haphazard previous behavior, output should be much more
predictable and useful.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>