topgit_deps.awk: omit annihilated self deps
If a branch is annihilated (and withan is false) then withbr
should not cause a self dependency edge to be omitted.

That's the whole point of "annihilated" branches -- to make
them disappear without having to edit all the .topdeps files.

Properly test for this and avoid omitting annihilated withbr
lines.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>