topgit_recurse.awk: always remove files when requested
Any of brfile, anfile, hdfile or rtfile may, perhaps,
remain unread in some less common circumstances.

Make sure that they are *always* removed when requested
(rmbr for brfile, rman for anfile, rmhd for hdfile and
rmrt for rtfile) and the script exits without any error.

Specifically this change addresses failures to remove when
either there was no input or possibly the file was not
processed due to option settings.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>