ref_prefixes.awk: minor efficiency
A sentinel array has "1" stored in it for each value that
it keeps track of having seen (i.e. `track[val] = 1`).

Instead of checking for a value with `track[val]`, check for
it with `(val in track)` to avoid growing the array when
it's not necessary.

It's only a minor efficiency, but every little bit helps.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>