topgit_recurse.awk: make multib mode ignore dupes
As described in the included docs, multib mode is supposed to
ignore duplicate starting branch names included in startb.

Unfortunately due to a typo it was not.

Correct this and enhance the semantics so that the prior
multi multib behavior remains available by setting multib
to an integer value greater than 1.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>