topgit_navigate.awk: always remove files when requested
Although either or both of anfile and brfile may remain
unread due to various specific combinations of input and/or
options, make sure that they are *always* removed when
requested (rman for anfile and rmbr for brfile) and the
script exits without any error.

Specifically this change addresses failures to remove when
either there was no input or possibly the file was not
processed due to option settings.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>