tg-info: make --leaves "describe" lightweight --tags
The documentation for `tg info --leaves` indicates that
"a tag name will be used when appropriate."  It does *NOT*
say only an _annotated_ tag will be used.

Add a fallback so that if an exact match using only annotated
tags cannot be found, another match is attempted using *ALL*
tags (i.e. including "lightweight" tags).

By making two attempts, the result will be guaranteed to prefer
an annotated tag name over a "lightweight" tag name no matter
what changes may or may not be made to the underlying logic in
the Git commands.  (Git ought to always prefer an annotated
tag over a lightweight tag when they both point to the same
commit, but it doesn't hurt to be extra careful just in case.)

Also update the t4300 test to reflect the new more correct
behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>