topgit_navigate.awk: count the step off of nil
The first step off of a nil ("") `startb` either goes to the
roots (forward) or heads (backward).  Moving as far as possible
(`steps=-1`) also goes to the heads (forward) or roots (backward).

Since the "as far as possible" off of nil has been optimized, the
code automatically converts a single step off of nil to an "as far
as possible" off of nil in the opposite direction.

Therefore as far as the code is concerned, a single step off of
nil never really takes place.

However, when `steps` is greater than 1 it's a different story.

The first step still moves to the roots (forward) or heads (backward),
but is NOT optimized in this case.

Thereafter the next step(s) operate(s) as expected.  Or at least they
are supposed to.

In this edge case where `startb=""` and `steps` is greater than 1,
the first step was not being counted, causing an extra step too far.

Correct this by counting the initial step off of nil.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>