topgit_navigate.awk: roots not always equal to starts
Consider the following TopGit DAG:

    C      # .topdeps file for TopGit branch "C"
   / \     A
  A   B    B

The above when linearized into a patch series produces:

  patch A
  patch B
  patch C

The "heads" of the TopGit DAG in this case is the single
"head" node "C".

The "roots" of the TopGit DAG in this case are the two
leaf nodes "A" and "B".

The "starting/first patches" of the represented patch series
in this case is the single patch/node A.

The "ending/last patches" of the represented patch series
in this case is the single patch/node C.

Because patch linearization enumerates the TopGit DAG in a
postfix order, the "heads" of the TopGit DAG will always
exactly correspond to the "ending/last patches".  Admittedly
the "last patch" could be an empty patch (aka a patch that
introduces no changes) that gets skipped on output, but it's
still actually there.

However, the "roots/leaves" of the TopGit DAG (those nodes
in the DAG that do not have any child nodes) do NOT always
exactly match the list of "starting/first patches"!

Therefore update the code to acknowledge this fact like so:

  === stepping from nil startb ===

              | forced |
  steps | rev | chklps | result
  ------|-----|--------|------------------------------------------------
    1   |  0  |   No   | "roots"/"leaves" of TopGit DAG
    1   |  1  |   No   | "heads"/"ending/final patches of patch series"
   -1   |  0  |   No   | "heads"/"ending/final patches of patch series"
   -1   |  1  |   Yes  | "starting/first patches of patch series"

The last row of the table represents a change in behavior.  Previously
a steps=-1 and rev=1 would also produce the same output as steps=1 rev=0
did (the first row of the table).

Not anymore.  While it was not previously possible to navigate to
all the first patches of all the series present at once (unless all
the TopGit nodes had at most one non-annihilated child), it is now.

As a result of the change, loop checking is now forced for the last
row of the table because the shortcut that obviates the need for
loop checking is not available with the new semantics for that row.

Update the comments with copious additions to explain all this too.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>