topgit_navigate.awk: after prune recalculate incoming and outgoing
After processing a requested graph prune operation (via the pruneb
option), recompute the incoming and outgoing links arrays.

When using inclbr/exclbr, this pruning takes place as the edges
are being read in, causing the incoming and outgoing arrays to
be constructed with the correct values from the beginning.

Since the graph must be navigated when performing a pruneb operation,
it's not possible to construct the final incoming and outgoing array
contents until later as the graph cannot be navigated until after
all the edges have been read in at which point the incoming and
outgoing arrays have already been built.

After severing graph segments as directed by the pruneb option,
simply iterate over the remaining outgoing edges to recreate the
correct resulting incoming and outgoing arrays for the new graph.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>