create-html.sh: minor nit use double-quotes
While html/xhtml tags allow attribute values to be either single-quoted
or double-quoted, in the case of the "meta" tag and its "http-equiv"
and "content" values it's more complex.

This "meta" tag:

    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

Creates a virtual HTTP header like this:

    Content-Type: text/html; charset=utf-8

Now that we are in "MIME" space, the attributes can only be
double-quoted.

For example this has identical semantics:

    Content-Type: text/html; charset="utf-8"

But this is invalid:

    Content-Type: text/html; charset='utf-8'

Update the generation of the "meta" "Refresh" virtual HTTP header
so that it's produced with double-quotes instead of single quotes.

In practice, consumers seem to be tolerating the single-quoted MIME
attribute value, but since it's not standards-compliant, change it.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>