t/t1200/maketable: replace perl version with awk version
Although the maketable script does not actually get run during the
t1200 test, it's provided in order to rebuild the table embedded
inside the t1200 script if necessary.

Eliminate the dependency on perl and replace the generating script
with an equivalent awk one.

It would be possible to replace the generator with a POSIX shell
script too (much slower run time though) if ever the necessity
arises.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>