tg.sh: do not pass arguments to dot command
The "dot" command does not accept arguments.  Attempting to
pass arguments to it could potentially cause an error
when using some theoretical shell that fails in such a case.

However, when using the "page" function to run a "dot"
command, the arguments MUST be passed to the page function
because it has its own scope for "$@" and they will disappear
otherwise since the "dot" command will be run within the
scope of the "page" function.

Add a "kludge" fix and check the command passed to the
"page" function.  If it starts with ". " then do NOT pass
along the "$@" arguments, otherwise do.

That pretty much handles everything that needs to be done
automatically so that clients of the "page" function need
not worry about it and just can always pass in the needed
arguments.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>