README: rename to README_DOCS.rst
The README really contains *ALL* of the documentation and is much
more than just a "Read Me" file.

Rename it to "README_DOCS.rst" to reflect that fact and the fact
that it's kept in ".rst" format.

Update all the necessary bits to look for it in the new location.

In preparation for adding a real "Read Me", remove the "README.rst"
link and replace it with a "README.md" link.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>