tg.sh: make .topdeps/.topmsg cache write atomic
When caching the contents of a .topmsg and/or .topdeps file in
the topgit cache area, make sure that it's either all cached or
all not cached.

Instead of writing each line to the actual cache file itself,
write to a temporary file name and then while the entire cache
file has been written, move it into place atomically.

This avoids the unfortunate situation where topgit is interrupted
while creating the cache for one of these files and it's only
half written.

With this change, if topgit is interrupted while writing the
cache file for .topmsg or .topdeps then nothing will be cached
at all rather than incorrect partial contents.

It's very difficult to trigger this problem in practice, but
with this change it's now no longer possible to trigger it at all.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>