tg--merging.sh: accomodate Git v2.30.0 checkout-index breakage
Since Git commit 7e41061588eab25d ("checkout-index: propagate errors to
exit code", 2020-10-27, v2.30.0), the `git checkout-index` command has
started exiting with a non-zero exit code when it encounters an argument
that does not exist in the index.

Even when the `-q/--quiet` option that is supposed to "be quiet if files
exist or are not in the index" has been given, it still produces an error.

Accomodate this breakage by using `git ls-files` first to determine
which of the TopGit files (.topdeps and/or .topmsg) are actually present
(if any) in the index and need to be forcibly checked out and then only
pass those names to the `checkout-index` command.

The `ls-files` command *does* ignore any arguments that do not match any
entries in the index and *does* exit without any error code on such
non-matches unless the explicit `--error-unmatch` option has been given.

While it would be possible to just ignore the exit code of `checkout-index`
instead (the code would be simpler), that would mean that any other errors
that were unrelated to a non-matching argument would also be ignored and
that's highly undesirable.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>