README: add topgit.top-bases setting to example
The heads location really is the more compatible location and
provides a better experience, but the default has not flipped
(and won't until the 0.20 release).

Add an extra `git config topgit.top-bases heads` line along with
a comment to the example to make it work as-is with the current
TopGit release.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>