awk/ref_match.awk: do not choke on '&' in ref names
Surprisingly, the '&' character does not cause Git to barf when it
occurs within a ref name.  In other words, this is okay:

    git update-ref refs/heads/surprise\& HEAD

(Presuming, of course, that HEAD is currently valid.)

Unfortunately, when expanding the '%(refname)' pattern, the actual
refname was appearing as the "repl" argument to awk's `gsub` function.
If that refname happened to contain an `&` character, the results
would be unexpected.

Correct the problem by properly escaping any `&` characters that may
occur inside the refname before using it as the "repl" argument.

Although potentially any `\` characters could also cause problems,
they are not allowed within ref names by Git.  And what the standard
says about the `\\` sequence within the "repl" argument doesn't
actually match the observed behavior of awk which makes it best to
leave any backslash characters alone to avoid introducing new issues.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>