awk/ref_match.awk: always sort refs without trailing '/'
When sorting ref names and ref name patterns, always make sure the
sorting takes place on ref names and ref name patterns that have
had their trailing '/' removed.

Git allows a number of characters with an ASCII value less than
that of the '/' character to be present in valid ref names.

Sorting with the '/' on the end (implied or explicit) causes refs
like this:

    refs/heads/a
    refs/heads/a-a

To end up being sorted incorrectly like this:

    refs/heads/a-a
    refs/heads/a

That's because the implied trailing '/' (ASCII 0x2F) comes after
the '-' (ASCII 0x2D).

Move a few things around so that matching patterns are sorted without
their trailing '/' but then have it added as they're being collected.

Make ref names be sorted and collected without their trailing '/'
but have it magically be added back to a temporary variable in the
filtering loop to make the filtering logic work properly.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>