t/Makefile.mak: add trap on INT for non-prove aggregate results
When running the tests without using Perl's `prove` utility, it's
nice to see summarized aggregate results of everything that's run
so far even if the testing process ends abnormally with an unexpected
error or via a Ctrl-C.

Add the missing trap on INT in the aggregate-results-and-cleanup
rule that should have been there from the beginning to make sure
the aggregate results get shown.

The check for a 'Bail out!', in which case no attempt is made to show
results ('Bail out!'s are very bad), remains in place.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>