t/test-lib-{main,functions}.sh: prohibit perl in tests
Do not allow perl to be used by the tests themselves.  Perl is
no longer required by the tests in the test suite.

However, perl continues to be required to properly format the
plain text documentation and to run test-related items such as
check-non-portable-shell.pl.

In other words, use of Perl is not going away completely any
time soon and likely never will.

None the less, having the tests themselves run without using Perl
provides additional stability for the test results themselves.

To that end, produce a fatal error if any test in the test suite
attempts to run perl.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>