test-lib*.sh: remove the last use of PERL_PATH
Reports of Perl's demise were premature.  There was one remaining
use of PERL_PATH in the test suite itself.

Remove it from test-lib.sh and replace it with the equivalent from
test-lib-main.sh with a few improvments.

At the same time tidy up the test-lib-main.sh copy to match the new
test-lib.sh version to hopefully make it more self-documenting.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>