test framework: add underpinnings for Git sha256 support
Add new environment variable TESTLIB_GIT_DEFAULT_HASH that permits
the default test hash algorithm to be changed globally for all the
tests.

Add new test variable test_hash_algo that permits tests to examine
the current test hash algorithm in use.

Add new individual test option --hash=<hash> that allows an individual
test to be run with the selected hash algorithm.

Add new prerequisites GITSHA1 and GITSHA256 that test, respectively,
whether sha1 is the currently active hash algorithm and whether or not
the version of Git being used for testing supports the sha256 algorithm.

Update the git_init test function to pass --object-format=<hash> as
appropriate for the current hash algorithm and Git version in use.

Add new test functions:

  test_asv
  test_asv_cache
  test_set_hash_algo
  test_v_asv
  test_v_git_mt

Eliminate the global variables EMPTY_TREE, EMPTY_BLOB and _z40.

Set and export GIT_DEFAULT_HASH to reflect the currently selected
hash algorithm.

Finally make a minor change to t0002-testlib-basic.sh to avoid having
it depend on the now-removed EMPTY_TREE variable.

The focus of these changes is providing the necessary underpinnings in
the test framework to support testing with a Git hash algorithm of sha256.

These changes are necessary, but not sufficient as they do not include
the additional changes required by some of the tests for sha256 support.

However, with the minor update to t0002, these changes do not affect
testing with hash algorithm sha1.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>