test-lib-*.sh: make sure GIT_DEFAULT_HASH gets into the test cache
When running multiple tests via the Makefile, the result of generic
test setup, which is the same for every test, is cached in a
TG-TEST-CACHE file.

In order for TESTLIB_GIT_DEFAULT_HASH to be effective in that
environment, the result of its effect on GIT_DEFAULT_HASH must not
be discarded and instead must be placed into TG-TEST-CACHE.

Adjust the caching logic to include GIT_DEFAULT_HASH in the test
cache whenever it's being used.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>