tg: initial Git sha256 hash algorithm support
The primary issue that arises when supporting a different hash
algorithm is that the "constant" hash values used for different
items such as the "null" object and the "empty" blob and tree all
change.

Update the code to automatically use the correct "constant" hash
values for the hash algorithm that's currently in use.

Also make sure the wayback machine preserves the correct hash
algorithm when activating wayback mode.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>