tg: improve Git sha256 hash algorithm support
In a number of cases it's necessary to match against a full hash
value.  The "$octet20" pattern has been used to do this for sha1.

Add an "$octet32" pattern and break down the basic "$octet" pattern
into two "$hexch" patterns to provide a foundation for sha256.

After determining which hash algorithm is in effect for the current
repository, set a new "$octethl" variable to either "$octet20" or
"$octet32" based on the active hash algorithm's digest length.

Use the new "$octethl" in place of "$octet20" where an exact hash
digest length is required.  In other cases where the source is
guaranteed to be a hash digest of the correct length, such as the
output of the `git ls-files --cached -s` command, it's possible to
be a bit sloppier and just use "$octet20$hexchar*" instead.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>