t6100: expunge test_tolerate_failure items
Split off the part of the two tests that simply cannot be performed
anymore starting with Git version 2.31.0 into their own tests and
make them have a prerequisite of NOT Git 2.31.0 or later.

Change the test_tolerate_failure back to test_expect_success on
the remaining parts of the original tests.

Add some explicit file checks to the original tests to replace the
bits that can no longer be done with Git 2.31.0 or later.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>