tg-update.sh: issue more messages when post-update checkout fails
Although tg-update ensures the worktree is clean before beginning
the update, that does not preclude the existence of untracked files.

When doing the final post-update checkout, if the new tree being
checked out contains a file that is already present as an untracked
file, then the checkout will fail.

Add an informational message when this happens.  If HEAD has
been left detached, add even more messages mentioning this as well
as the exact git checkout command to use in that case (it will still
fail, but at least that provides a starting point).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>